move frienship queries
