+++ ChainLink-Client/src/graphql/queries/userQueries.ts
@@ -25,26 +25,6 @@ export const FETCH_USER_BY_NAME = gql`
   }
 `;
 
-export const GET_FRIENDSHIP_STATUS = gql`
-  query getFriendshipStatus($sender: String!, $receiver: String!) {
-    getFriendshipStatus(sender: $sender, receiver: $receiver) {
-      status
-      }
-  }
-`;
-
-export const GET_FRIENDS = gql`
-  query getFriends($username: String!) {
-    getFriends(username: $username)
-  }
-`;
-
-export const GET_FRIEND_REQUESTS = gql`
-  query getFriendRequests($username: String!) {
-    getFriendRequests(username: $username)
-  }
-`;
-
 export const VALIDATE_USERNAME = gql`
   query validUsername($username: String!) {
     validUsername(username: $username)
